refactor: move the CRS code to its own module and drop the planar projection subclass
CrsDefinition, CrsWorldCoordinateHelper and simpleCrs live in crs.ts. The
simple CRS is a class with one exported instance. PlanarProjection only
carried a name over MercatorProjection, so MercatorProjection takes the
name instead and the subclass is gone.
