+++ src/geo/projection/mercator_transform.test.ts
@@ -11,7 +11,7 @@ import {createDEM, createDEMTerrain, createTerrain, expectToBeCloseToArray} from
 import {EXTENT} from '../../data/extent.ts';
 import {MercatorCoordinate} from '../mercator_coordinate.ts';
 import type {Tile} from '../../tile/tile.ts';
-import {CrsWorldCoordinateHelper, simpleCrs} from './planar_projection.ts';
+import {CrsWorldCoordinateHelper, simpleCrs} from './crs.ts';
 
 describe('transform', () => {
     test('creates a transform', () => {
