+++ src/index.ts
@@ -67,7 +67,7 @@ import type {RequestTransformFunction, ResourceType} from './util/request_manage
 import type {OverscaledTileID} from './tile/tile_id.ts';
 import type {PositionAnchor} from './ui/anchor.ts';
 import type {CustomLayerProjectionData, ProjectionData, ProjectionDataParams, ProjectionMatrix, RendererProjectionData} from './geo/projection/projection_data.ts';
-import type {CrsDefinition} from './geo/projection/planar_projection.ts';
+import type {CrsDefinition} from './geo/projection/crs.ts';
 import type {TileMatrix} from './geo/projection/tile_matrix.ts';
 import type {WorkerTileResult} from './source/worker_source.ts';
 import type {Actor, IActor} from './util/actor.ts';
